Imporve casts to compile without errors on linux gcc

+++ src/ptmf.cpp
@@ -8,6 +8,7 @@
 
 /* -- Includes ------------------------------------------------------------ */
 #include <string.h>
+#include <stdint.h>
 #include <stdio.h>
 #include "ptmf.h"
 
@@ -28,11 +29,11 @@ void PtmfA::learnA()
 {
    printf("__________PtmfA__________\n");
    printf("this => %p\n", this);
-   printf("sizeof(PtmfA) => %d\n", sizeof(PtmfA));
+   printf("sizeof(PtmfA) => %d\n", (int)sizeof(PtmfA));
    printf("----------DATA-----------\n");
-   printf("&this->a1 => %p (this + %d)\n", &this->a1, ((int)&this->a1 - (int)this));
-   printf("&this->a2 => %p (this + %d)\n", &this->a2, ((int)&this->a2 - (int)this));
-   printf("&this->a3 => %p (this + %d)\n", &this->a3, ((int)&this->a3 - (int)this));
+   printf("&this->a1 => %p (this + %d)\n", &this->a1, (int)((uintptr_t)&this->a1 - (uintptr_t)this));
+   printf("&this->a2 => %p (this + %d)\n", &this->a2, (int)((uintptr_t)&this->a2 - (uintptr_t)this));
+   printf("&this->a3 => %p (this + %d)\n", &this->a3, (int)((uintptr_t)&this->a3 - (uintptr_t)this));
    printf("----------PTMF-----------\n");
 
    //i need the union to examine the pointer to member function ...
@@ -51,11 +52,11 @@ void PtmfB::learnB()
 {
    printf("__________PtmfB__________\n");
    printf("this => %p\n", this);
-   printf("sizeof(PtmfA) => %d\n", sizeof(PtmfB));
+   printf("sizeof(PtmfA) => %d\n", (int)sizeof(PtmfB));
    printf("----------DATA-----------\n");
-   printf("&this->b1 => %p (this + %d)\n", &this->b1, ((int)&this->b1 - (int)this));
-   printf("&this->b2 => %p (this + %d)\n", &this->b2, ((int)&this->b2 - (int)this));
-   printf("&this->b3 => %p (this + %d)\n", &this->b3, ((int)&this->b3 - (int)this));
+   printf("&this->b1 => %p (this + %d)\n", &this->b1, (int)((uintptr_t)&this->b1 - (uintptr_t)this));
+   printf("&this->b2 => %p (this + %d)\n", &this->b2, (int)((uintptr_t)&this->b2 - (uintptr_t)this));
+   printf("&this->b3 => %p (this + %d)\n", &this->b3, (int)((uintptr_t)&this->b3 - (uintptr_t)this));
    printf("----------PTMF-----------\n");
 
    //i need the union to examine the pointer to member function ...
@@ -81,11 +82,11 @@ void PtmfC::learnC()
 {
    printf("__________PtmfC__________\n");
    printf("this => %p\n", this);
-   printf("sizeof(PtmfA) => %d\n", sizeof(PtmfC));
+   printf("sizeof(PtmfA) => %d\n", (int)sizeof(PtmfC));
    printf("----------DATA-----------\n");
-   printf("&this->c1 => %p (this + %d)\n", &this->c1,  ((int)&this->c1  - (int)this));
-   printf("&this->c2 => %p (this + %d)\n", &this->c2,  ((int)&this->c2  - (int)this));
-   printf("&this->c3 => %p (this + %d)\n", &this->c3,  ((int)&this->c3  - (int)this));
+   printf("&this->c1 => %p (this + %d)\n", &this->c1, (int)((uintptr_t)&this->c1 - (uintptr_t)this));
+   printf("&this->c2 => %p (this + %d)\n", &this->c2, (int)((uintptr_t)&this->c2 - (uintptr_t)this));
+   printf("&this->c3 => %p (this + %d)\n", &this->c3, (int)((uintptr_t)&this->c3 - (uintptr_t)this));
    printf("----------PTMF-----------\n");
 
    //i need the union to examine the pointer to member function ...
@@ -142,24 +143,24 @@ void PtmfX::learnX()
 {
    printf("__________PtmfX__________\n");
    printf("this => %p\n", this);
-   printf("sizeof(PtmfX) => %d\n", sizeof(PtmfX));
+   printf("sizeof(PtmfX) => %d\n", (int)sizeof(PtmfX));
    printf("----------OFFSETS-----------\n");
-   printf("(PtmfA *)this => %p (this + %d)\n", (PtmfA *)this, ((int)((PtmfA *)this) - (int)this));
-   printf("(PtmfB *)this => %p (this + %d)\n", (PtmfB *)this, ((int)((PtmfB *)this) - (int)this));
-   printf("(PtmfC *)this => %p (this + %d)\n", (PtmfC *)this, ((int)((PtmfC *)this) - (int)this));
+   printf("(PtmfA *)this => %p (this + %d)\n", (PtmfA *)this, (int)((uintptr_t)((PtmfA *)this) - (uintptr_t)this));
+   printf("(PtmfB *)this => %p (this + %d)\n", (PtmfB *)this, (int)((uintptr_t)((PtmfB *)this) - (uintptr_t)this));
+   printf("(PtmfC *)this => %p (this + %d)\n", (PtmfC *)this, (int)((uintptr_t)((PtmfC *)this) - (uintptr_t)this));
    printf("----------DATA-----------\n");
-   printf("&this->a1 => %p (this + %d)\n", &this->a1, ((int)&this->a1 - (int)this));
-   printf("&this->a2 => %p (this + %d)\n", &this->a2, ((int)&this->a2 - (int)this));
-   printf("&this->a3 => %p (this + %d)\n", &this->a3, ((int)&this->a3 - (int)this));
-   printf("&this->b1 => %p (this + %d)\n", &this->b1, ((int)&this->b1 - (int)this));
-   printf("&this->b2 => %p (this + %d)\n", &this->b2, ((int)&this->b2 - (int)this));
-   printf("&this->b3 => %p (this + %d)\n", &this->b3, ((int)&this->b3 - (int)this));
-   printf("&this->c1 => %p (this + %d)\n", &this->c1, ((int)&this->c1 - (int)this));
-   printf("&this->c2 => %p (this + %d)\n", &this->c2, ((int)&this->c2 - (int)this));
-   printf("&this->c3 => %p (this + %d)\n", &this->c3, ((int)&this->c3 - (int)this));
-   printf("&this->x1 => %p (this + %d)\n", &this->x1, ((int)&this->x1 - (int)this));
-   printf("&this->x2 => %p (this + %d)\n", &this->x2, ((int)&this->x2 - (int)this));
-   printf("&this->x3 => %p (this + %d)\n", &this->x3, ((int)&this->x3 - (int)this));
+   printf("&this->a1 => %p (this + %d)\n", &this->a1, (int)((uintptr_t)&this->a1 - (uintptr_t)this));
+   printf("&this->a2 => %p (this + %d)\n", &this->a2, (int)((uintptr_t)&this->a2 - (uintptr_t)this));
+   printf("&this->a3 => %p (this + %d)\n", &this->a3, (int)((uintptr_t)&this->a3 - (uintptr_t)this));
+   printf("&this->b1 => %p (this + %d)\n", &this->b1, (int)((uintptr_t)&this->b1 - (uintptr_t)this));
+   printf("&this->b2 => %p (this + %d)\n", &this->b2, (int)((uintptr_t)&this->b2 - (uintptr_t)this));
+   printf("&this->b3 => %p (this + %d)\n", &this->b3, (int)((uintptr_t)&this->b3 - (uintptr_t)this));
+   printf("&this->c1 => %p (this + %d)\n", &this->c1, (int)((uintptr_t)&this->c1 - (uintptr_t)this));
+   printf("&this->c2 => %p (this + %d)\n", &this->c2, (int)((uintptr_t)&this->c2 - (uintptr_t)this));
+   printf("&this->c3 => %p (this + %d)\n", &this->c3, (int)((uintptr_t)&this->c3 - (uintptr_t)this));
+   printf("&this->x1 => %p (this + %d)\n", &this->x1, (int)((uintptr_t)&this->x1 - (uintptr_t)this));
+   printf("&this->x2 => %p (this + %d)\n", &this->x2, (int)((uintptr_t)&this->x2 - (uintptr_t)this));
+   printf("&this->x3 => %p (this + %d)\n", &this->x3, (int)((uintptr_t)&this->x3 - (uintptr_t)this));
    printf("----------CODE-----------\n");
 
    //i need the union to examine the pointer to member function ...
